Classes and interfaces refactoring
