+++ MicroBlog/src/main/java/it/gfurri20/blog/domain/auth/RegistrationModelView.java
@@ -1,0 +1,17 @@
+
+package it.gfurri20.blog.domain.auth;
+
+import lombok.Data;
+
+
+/**
+ *
+ * @author gfurri20
+ */
+@Data
+public class RegistrationModelView
+{
+    private String username;
+    private String password;
+    private String repeatPassword;
+}
