+++ MicroBlog/src/main/java/it/gfurri20/blog/service/impl/BlogCommentService.java
@@ -1,8 +1,11 @@
 
-package it.gfurri20.blog.service;
+package it.gfurri20.blog.service.impl;
 
 import it.gfurri20.blog.domain.BlogComment;
 import it.gfurri20.blog.repository.IBlogCommentRepository;
+import it.gfurri20.blog.service.IBlogCommentService;
+import it.gfurri20.blog.service.IBlogPostService;
+import it.gfurri20.blog.service.IBlogUserService;
 import java.util.Date;
 import java.util.List;
 import java.util.NoSuchElementException;
